+++ server/provider/provider.go
@@ -22,13 +22,12 @@ import (
 	"fmt"
 	"sync"
 
-	"github.com/nymtech/nym/server"
-
-	coconut "github.com/nymtech/nym/crypto/coconut/scheme"
-	"github.com/nymtech/nym/logger"
-	"github.com/nymtech/nym/server/config"
 	ethcrypto "github.com/ethereum/go-ethereum/crypto"
 	Curve "github.com/jstuczyn/amcl/version3/go/amcl/BLS381"
+	coconut "github.com/nymtech/nym/crypto/coconut/scheme"
+	"github.com/nymtech/nym/logger"
+	"github.com/nymtech/nym/server"
+	"github.com/nymtech/nym/server/config"
 	"gopkg.in/op/go-logging.v1"
 )