+++ tendermint/nymabci/nymapplication/deliver.go
@@ -17,15 +17,16 @@
 package nymapplication
 
 import (
+	"encoding/base64"
+	"encoding/binary"
 	"fmt"
 
+	ethcommon "github.com/ethereum/go-ethereum/common"
+	"github.com/golang/protobuf/proto"
 	"github.com/nymtech/nym/constants"
-	coconut "github.com/nymtech/nym/crypto/coconut/scheme"
 	"github.com/nymtech/nym/tendermint/nymabci/code"
 	tmconst "github.com/nymtech/nym/tendermint/nymabci/constants"
 	"github.com/nymtech/nym/tendermint/nymabci/transaction"
-	ethcommon "github.com/ethereum/go-ethereum/common"
-	"github.com/golang/protobuf/proto"
 	"github.com/tendermint/tendermint/abci/types"
 	cmn "github.com/tendermint/tendermint/libs/common"
 )
@@ -182,15 +183,6 @@ func (app *NymApplication) handleDepositCredential(reqb []byte) types.ResponseDe
 		return types.ResponseDeliverTx{Code: checkResult}
 	}
 
-	// the errors were checked at checkDepositCoconutCredentialTx call, so they can't by anything else but nil
-	// if it's not the case, we can't trust anything that is happening anyway so we can only panic
-	cred := &coconut.Signature{}
-	theta := &coconut.ThetaTumbler{}
-	pubM, err := coconut.BigSliceFromByteSlices(req.PubM)
-	mustNilErr(err)
-	mustNilErr(cred.FromProto(req.Sig))
-	mustNilErr(theta.FromProto(req.Theta))
-
 	address := ethcommon.BytesToAddress(req.ProviderAddress)
 
 	if !app.checkIfAccountExists(address[:]) {
@@ -203,30 +195,32 @@ func (app *NymApplication) handleDepositCredential(reqb []byte) types.ResponseDe
 		}
 		app.log.Debug(fmt.Sprintf("Created new account for %v", address.Hex()))
 	}
-	//
-	// Once verification is moved to separate entity, the below will be used
-	//
-	// protoSigB, err := proto.Marshal(req.Sig)
-	// if err != nil {
-	// 	app.log.Error("Failed to marshal the received credential")
-	// 	return types.ResponseDeliverTx{Code: code.INVALID_TX_PARAMS}
-	// }
-	// protoThetaB, err := proto.Marshal(req.Theta)
-	// if err != nil {
-	// 	app.log.Error("Failed to marshal the received crypto materials")
-	// 	return types.ResponseDeliverTx{Code: code.INVALID_TX_PARAMS}
-	// }
-	// key := make([]byte, ethcommon.AddressLength + len(protoSigB) + len(tmconst.RedeemTokensRequestKeyPrefix))
-	// i := copy(key, tmconst.RedeemTokensRequestKeyPrefix)
-	// i += copy(key[i:], address)
-	// copy(key[i:], protoSigB)
-	// return types.ResponseDeliverTx{
-	// 	Code: code.OK,
-	// 	Tags: []cmn.KVPair{
-	// 		// [ Prefix || Provider || credential --- required crypto materials ]
-	// 		{Key: key, Value: protoThetaB},
-	// 	},
-	// }
+
+	cryptoMaterialsBytes, err := proto.Marshal(req.CryptoMaterials)
+	if err != nil {
+		return types.ResponseDeliverTx{Code: code.INVALID_TX_PARAMS}
+	}
+
+	app.log.Debug(fmt.Sprintf("Deposit request from address %v, zeta %v", req.ProviderAddress, req.CryptoMaterials.Theta.Zeta))
+
+	app.setZetaStatus(req.CryptoMaterials.Theta.Zeta, tmconst.ZetaStatusBeingVerified)
+
+	key := make([]byte, ethcommon.AddressLength+len(req.CryptoMaterials.Theta.Zeta)+len(tmconst.RedeemTokensRequestKeyPrefix)+8)
+	i := copy(key, tmconst.RedeemTokensRequestKeyPrefix)
+	i += copy(key[i:], address[:])
+	binary.BigEndian.PutUint64(key[i:], uint64(req.Value))
+	i += 8
+	copy(key[i:], req.CryptoMaterials.Theta.Zeta)
+	return types.ResponseDeliverTx{
+		Code: code.OK,
+		Tags: []cmn.KVPair{
+			// while it is not crucial we have unique keys here, verifiers will need to be able to
+			// send a transaction back "confirming" status of this data and this will require an unique key field.
+			// So we might as well use the same system already
+			// [ Prefix || Provider || uint64(value) || Zeta(g^s) --- required crypto materials ]
+			{Key: key, Value: cryptoMaterialsBytes},
+		},
+	}
 
 	// everything below this line will be moved to separate entity (in a way) it will be replaced by the commneted
 	// code above
@@ -237,35 +231,97 @@ func (app *NymApplication) handleDepositCredential(reqb []byte) types.ResponseDe
 	//
 	// TODO: credential and proof verification will be moved to another 'verifier' entity
 	// but for test sake, let's just leave them here for a time being.
-	avk, err := app.retrieveAggregateVerificationKey()
-	if err != nil {
-		app.log.Error("Failed to retrieve verification key")
-		return types.ResponseDeliverTx{Code: code.UNKNOWN}
+	// avk, err := app.retrieveAggregateVerificationKey()
+	// if err != nil {
+	// 	app.log.Error("Failed to retrieve verification key")
+	// 	return types.ResponseDeliverTx{Code: code.UNKNOWN}
+	// }
+
+	// // NOTE: TODO:
+	// // if credentials were to be verified during delivertx rather than by separate entity, there's no
+	// // point in generating those params every deliverTx. Just store them in state and generate them every time
+	// // server restarts (or they are nil)
+	// params := app.getSimpleCoconutParams()
+	// if params == nil {
+	// 	app.log.Error("Failed to generate coconut params")
+	// 	return types.ResponseDeliverTx{Code: code.UNKNOWN}
+	// }
+	// // verify the credential
+	// isValid := coconut.BlindVerifyTumbler(params, avk, cred, theta, pubM, address[:])
+
+	// if isValid {
+	// 	app.log.Debug("The received credential was valid")
+	// 	if err := app.increaseBalanceBy(address[:], uint64(req.Value)); err != nil {
+	// 		app.log.Error("failed to increase provider's balance? Critical failure")
+	// 		panic(err)
+	// 	}
+	// 	// store the used credential
+	// 	app.storeSpentZeta(req.CryptoMaterials.Theta.Zeta)
+	// 	return types.ResponseDeliverTx{Code: code.OK}
+	// }
+
+	// app.log.Debug("The received credential was invalid")
+	// return types.ResponseDeliverTx{Code: code.INVALID_CREDENTIAL}
+}
+
+func (app *NymApplication) handleCredentialVerificationNotification(reqb []byte) types.ResponseDeliverTx {
+	req := &transaction.CredentialVerificationNotification{}
+
+	if err := proto.Unmarshal(reqb, req); err != nil {
+		return types.ResponseDeliverTx{Code: code.INVALID_TX_PARAMS}
 	}
 
-	// NOTE: TODO:
-	// if credentials were to be verified during delivertx rather than by separate entity, there's no
-	// point in generating those params every deliverTx. Just store them in state and generate them every time
-	// server restarts (or they are nil)
-	params := app.getSimpleCoconutParams()
-	if params == nil {
-		app.log.Error("Failed to generate coconut params")
-		return types.ResponseDeliverTx{Code: code.UNKNOWN}
+	if checkResult := app.checkCredentialVerificationNotification(reqb); checkResult != code.OK {
+		app.log.Info("handleCredentialVerificationNotification failed checkTx")
+		return types.ResponseDeliverTx{Code: checkResult}
 	}
-	// verify the credential
-	isValid := coconut.BlindVerifyTumbler(params, avk, cred, theta, pubM, address[:])
 
-	if isValid {
-		app.log.Debug("The received credential was valid")
-		if err := app.increaseBalanceBy(address[:], uint64(req.Value)); err != nil {
-			app.log.Error("failed to increase provider's balance? Critical failure")
-			panic(err)
+	// 'accept' the notification
+	newCount := app.storeVerifierNotification(req.VerifierPublicKey, req.Zeta, req.Value, req.CredentialValidity)
+	zetaB64 := base64.StdEncoding.EncodeToString(req.Zeta)
+
+	app.log.Debug(fmt.Sprintf("Reached %v notifications out of required %v for zeta %v (value %v)",
+		newCount,
+		app.state.verifierThreshold,
+		zetaB64,
+		req.Value,
+	))
+
+	// commit the transaction if threshold is reached
+	if newCount == app.state.verifierThreshold {
+		app.log.Debug(fmt.Sprintf("Reached required threshold of %v for %v (value %v)",
+			app.state.verifierThreshold,
+			zetaB64,
+			req.Value,
+		))
+
+		// check if account exists
+		currentBalance, err := app.retrieveAccountBalance(req.ProviderAddress)
+		// It should already exist since the provider had to send a request
+		// before to actually request the deposit to happen, but double check it now anyway
+		if err != nil && createAccountOnDepositIfDoesntExist {
+			didSucceed := app.createNewAccountOp(ethcommon.BytesToAddress(req.ProviderAddress))
+			if !didSucceed {
+				app.log.Info(fmt.Sprintf("Failed to create new account for the provider with address %v",
+					ethcommon.BytesToAddress(req.ProviderAddress).Hex()))
+				return types.ResponseDeliverTx{Code: code.UNKNOWN}
+			}
+		} else if err != nil {
+			app.log.Info("Provider's account does not exist and system is not set to create new ones")
+			return types.ResponseDeliverTx{Code: code.ACCOUNT_DOES_NOT_EXIST}
 		}
-		// store the used credential
-		app.storeSpentZeta(req.Theta.Zeta)
-		return types.ResponseDeliverTx{Code: code.OK}
+
+		app.setAccountBalance(req.ProviderAddress, currentBalance+uint64(req.Value))
+		// mark zeta as 'fully' spent
+		// TODO: perhaps do some mark and sweep later on for all credentials set as being verified for a long time
+		// and invalidate them?
+		app.log.Info(fmt.Sprintf("Marking zeta %v as spent and provider's %v balance was increased by %v",
+			zetaB64,
+			ethcommon.BytesToAddress(req.ProviderAddress).Hex(),
+			req.Value,
+		))
+		app.setZetaStatus(req.Zeta, tmconst.ZetaStatusSpent, req.ProviderAddress...)
 	}
 
-	app.log.Debug("The received credential was invalid")
-	return types.ResponseDeliverTx{Code: code.INVALID_CREDENTIAL}
+	return types.ResponseDeliverTx{Code: code.OK}
 }