+++ tendermint/nymabci/nymapplication/state.go
@@ -17,15 +17,16 @@
 package nymapplication
 
 import (
+	"bytes"
 	"encoding/base64"
 	"encoding/binary"
 	"errors"
 	"fmt"
 
-	coconut "github.com/nymtech/nym/crypto/coconut/scheme"
-	tmconst "github.com/nymtech/nym/tendermint/nymabci/constants"
 	ethcommon "github.com/ethereum/go-ethereum/common"
 	Curve "github.com/jstuczyn/amcl/version3/go/amcl/BLS381"
+	coconut "github.com/nymtech/nym/crypto/coconut/scheme"
+	tmconst "github.com/nymtech/nym/tendermint/nymabci/constants"
 	"github.com/tendermint/iavl"
 )
 
@@ -39,8 +40,9 @@ var (
 type State struct {
 	db *iavl.MutableTree // hash and height (version) are obtained from the tree methods
 
-	watcherThreshold uint32
-	pipeAccount      ethcommon.Address
+	watcherThreshold  uint32
+	verifierThreshold uint32
+	pipeAccount       ethcommon.Address
 }
 
 func (app *NymApplication) storeWatcherThreshold() {
@@ -56,6 +58,22 @@ func (app *NymApplication) loadWatcherThreshold() error {
 	}
 	app.state.watcherThreshold = binary.BigEndian.Uint32(val)
 	app.log.Info(fmt.Sprintf("Loaded watcher threshold: %v", app.state.watcherThreshold))
+	return nil
+}
+
+func (app *NymApplication) storeVerifierThreshold() {
+	thrb := make([]byte, 4)
+	binary.BigEndian.PutUint32(thrb, app.state.verifierThreshold)
+	app.state.db.Set(tmconst.VerifierThresholdKey, thrb)
+}
+
+func (app *NymApplication) loadVerifierThreshold() error {
+	_, val := app.state.db.Get(tmconst.VerifierThresholdKey)
+	if val == nil {
+		return ErrKeyDoesNotExist
+	}
+	app.state.verifierThreshold = binary.BigEndian.Uint32(val)
+	app.log.Info(fmt.Sprintf("Loaded verifier threshold: %v", app.state.verifierThreshold))
 	return nil
 }
 
@@ -164,6 +182,19 @@ func (app *NymApplication) checkWatcherKey(publicKey []byte) bool {
 	return app.state.db.Has(dbEntry)
 }
 
+func (app *NymApplication) storeVerifierKey(verifier Verifier) {
+	pubB64 := base64.StdEncoding.EncodeToString(verifier.PublicKey)
+	app.log.Debug(fmt.Sprintf("Adding to the trusted set verifier with public key: %v", pubB64))
+	dbEntry := prefixKey(tmconst.CredentialVerifierKeyPrefix, verifier.PublicKey)
+	// TODO: do we even need to set any meaningful value here?
+	app.state.db.Set(dbEntry, tmconst.CredentialVerifierKeyPrefix)
+}
+
+func (app *NymApplication) checkVerifierKey(publicKey []byte) bool {
+	dbEntry := prefixKey(tmconst.CredentialVerifierKeyPrefix, publicKey)
+	return app.state.db.Has(dbEntry)
+}
+
 // checks if given (random) nonce was already seen before for the particular address
 func (app *NymApplication) checkNonce(nonce, address []byte) bool {
 	if len(nonce) != tmconst.NonceLength || len(address) != ethcommon.AddressLength {
@@ -190,8 +221,8 @@ func (app *NymApplication) storeWatcherNotification(watcherKey, txHash []byte) u
 	// now increase notification count for this transaction
 	app.state.db.Set(key, tmconst.EthereumWatcherNotificationPrefix)
 	// and update total count
-	newCount := app.getNotificationCount(txHash) + 1
-	app.updateNotificationCount(txHash, newCount)
+	newCount := app.getPipeTransferNotificationCount(txHash) + 1
+	app.updatePipeTransferNotificationCount(txHash, newCount)
 	return newCount
 }
 
@@ -201,7 +232,7 @@ func (app *NymApplication) checkWatcherNotification(watcherKey, txHash []byte) b
 	return app.state.db.Has(key)
 }
 
-func (app *NymApplication) getNotificationCount(txHash []byte) uint32 {
+func (app *NymApplication) getPipeTransferNotificationCount(txHash []byte) uint32 {
 	key := prefixKey(tmconst.PipeAccountTransferNotificationCountKeyPrefix, txHash)
 
 	_, val := app.state.db.Get(key)
@@ -211,7 +242,7 @@ func (app *NymApplication) getNotificationCount(txHash []byte) uint32 {
 	return binary.BigEndian.Uint32(val)
 }
 
-func (app *NymApplication) updateNotificationCount(txHash []byte, count uint32) {
+func (app *NymApplication) updatePipeTransferNotificationCount(txHash []byte, count uint32) {
 	key := prefixKey(tmconst.PipeAccountTransferNotificationCountKeyPrefix, txHash)
 	countb := make([]byte, 4)
 	binary.BigEndian.PutUint32(countb, count)
@@ -219,13 +250,99 @@ func (app *NymApplication) updateNotificationCount(txHash []byte, count uint32) 
 	app.state.db.Set(key, countb)
 }
 
-func (app *NymApplication) storeSpentZeta(zeta []byte) {
-	key := prefixKey(tmconst.SpentZetaPrefix, zeta)
-	// TODO: the recurring problem of what value it should be set to.
-	app.state.db.Set(key, tmconst.SpentZetaPrefix)
+// additional data is used as a suffix for zeta status
+// currently only used for spent status to more easily query for whom redeemed given zeta
+func (app *NymApplication) setZetaStatus(zeta []byte, status tmconst.ZetaStatus, additionalData ...byte) {
+	key := prefixKey(tmconst.ZetaStatusPrefix, zeta)
+	value := status.DbEntry()
+	if len(additionalData) > 0 {
+		value = prefixKey(value, additionalData)
+	}
+	app.state.db.Set(key, value)
 }
 
-func (app *NymApplication) checkIfZetaIsSpent(zeta []byte) bool {
-	key := prefixKey(tmconst.SpentZetaPrefix, zeta)
+func (app *NymApplication) checkIfZetaIsUnspent(zeta []byte) bool {
+	key := prefixKey(tmconst.ZetaStatusPrefix, zeta)
+	_, status := app.state.db.Get(key)
+	if status == nil {
+		return true
+	}
+	return bytes.HasPrefix(status, tmconst.ZetaStatusSpent.DbEntry())
+}
+
+func (app *NymApplication) checkZetaStatus(zeta []byte) []byte {
+	key := prefixKey(tmconst.ZetaStatusPrefix, zeta)
+	_, status := app.state.db.Get(key)
+	if status == nil {
+		return tmconst.ZetaStatusUnspent.DbEntry()
+	}
+	return status
+}
+
+// returns new number of notifications received for this transaction
+// TODO: rethink if we need to include value in the key field or even at all here, because in principle
+// there can be no other credential of different value with the same zeta
+func (app *NymApplication) storeVerifierNotification(verifierKey, zeta []byte, value int64, valid bool) uint32 {
+	key := make([]byte, len(verifierKey)+len(zeta)+8)
+	i := copy(key, verifierKey)
+	i += copy(key[i:], zeta)
+	binary.BigEndian.PutUint64(key[i:], uint64(value))
+
+	// [PREFIX || VERIFIER || uint64(VALUE) || ZETA ]
+	key = prefixKey(tmconst.CredentialVerifierNotificationPrefix, key)
+
+	// again, does the value matter here? we could just set an empty array to save on space
+	// first store that given verifier sent the notification
+	app.state.db.Set(key, tmconst.CredentialVerifierNotificationPrefix)
+
+	// then update the global count, but only if credential is valid
+	currentCount := app.getCredentialVerificationCount(zeta, value)
+	if valid {
+		newCount := currentCount + 1
+		app.updateCredentialVerificationNotificationCount(zeta, value, newCount)
+		return newCount
+	}
+	return currentCount
+}
+
+// checks if this verifier has already sent notification regarding this credential
+func (app *NymApplication) checkVerifierNotification(verifierKey, zeta []byte, value int64) bool {
+	key := make([]byte, len(verifierKey)+len(zeta)+8)
+	i := copy(key, verifierKey)
+	i += copy(key[i:], zeta)
+	binary.BigEndian.PutUint64(key[i:], uint64(value))
+
+	// [PREFIX || VERIFIER || uint64(VALUE) || ZETA ]
+	key = prefixKey(tmconst.CredentialVerifierNotificationPrefix, key)
+
 	return app.state.db.Has(key)
+}
+
+func (app *NymApplication) getCredentialVerificationCount(zeta []byte, value int64) uint32 {
+	key := make([]byte, len(zeta)+8)
+	i := copy(key, zeta)
+	binary.BigEndian.PutUint64(key[i:], uint64(value))
+
+	// [PREFIX || uint64(VALUE) || ZETA ]
+	key = prefixKey(tmconst.CredentialVerificationNotificationCountKeyPrefix, key)
+
+	_, val := app.state.db.Get(key)
+	if val == nil {
+		return 0
+	}
+	return binary.BigEndian.Uint32(val)
+}
+
+func (app *NymApplication) updateCredentialVerificationNotificationCount(zeta []byte, value int64, count uint32) {
+	key := make([]byte, len(zeta)+8)
+	i := copy(key, zeta)
+	binary.BigEndian.PutUint64(key[i:], uint64(value))
+
+	// [PREFIX || uint64(VALUE) || ZETA ]
+	key = prefixKey(tmconst.CredentialVerificationNotificationCountKeyPrefix, key)
+
+	countb := make([]byte, 4)
+	binary.BigEndian.PutUint32(countb, count)
+
+	app.state.db.Set(key, countb)
 }